job-portal/backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Database
pymongo>=4.6.0
certifi>=2023.11.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Environment
python-dotenv>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Data Processing
python-dateutil>=2.8.2
geopy>=2.4.0

# Storage
minio>=7.2.0

# Rate Limiting
slowapi>=0.1.9

# HTTP Client
requests>=2.31.0
